// backend_vibeazy/src/interfaces/flutterwave-interface.ts
export interface WebhookDataProd {
  event: 'charge.completed',
  data: {
    id: number,
    tx_ref: string,
    reference: string
    flw_ref: string,
    device_fingerprint: string,
    amount: 30,
    currency: 'NGN',
    charged_amount: number,
    app_fee: number,
    merchant_fee: number,
    processor_response: string,
    auth_model:string,
    ip: string,
    narration:string,
    status: string,
    payment_type: string,
    created_at: string,
    account_id: 683269,
    customer: {
      id: 994192822,
      name: string,
      phone_number: string,
      email: string,
      created_at: string
    }
  },
  'event.type': 'BANK_TRANSFER_TRANSACTION' | "Transfer"
}